# WeatherStation.Api/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WeatherStation.Api, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/WeatherReadings": {
      "get": {
        "tags": [
          "WeatherReadings"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherReading"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherReading"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherReading"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "WeatherReadings"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeatherReading"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/WeatherReading"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/WeatherReading"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherReading"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/WeatherReadings/latest": {
      "get": {
        "tags": [
          "WeatherReadings"
        ],
        "parameters": [
          {
            "name": "location",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherReading"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherReading"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherReading"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "WeatherReading": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "temperatureC": {
            "type": "number",
            "format": "double"
          },
          "humidityPercent": {
            "type": "number",
            "format": "double"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      }
    }
  }
}